index: only collect metadata for stale paths
Stop collecting all of the metadata for every path, even if the path
hasn't changed since the last save.  Further, rework the code to
short-circuit some other unnecessary work.

To support this, split Entry.from_stat() into two parts, stale(), which
tests to see if an entry has "materially" changed, and
update_from_stat() which updates the entry to match the stat information
provided.

This should substantially decrease the indexing cost for paths that
haven't changed since the last save.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>